+++ app/javascript/users/pages/dashboard.vue
@@ -5,8 +5,8 @@
     <h3 v-if="course_events.length == 0 && !loading"
         class="d-flex justify-center subtitle">{{ $t('course_event.list.none') }}</h3>
 
-    <CourseEventsListCards v-if="course_events.length > 0" :course_events="course_events" />
-    <h3 v-else-if="course_events.length == 0 && !loading"
+    <CourseEventsListCards v-if="subscribed_course_events.length > 0" :course_events="subscribed_course_events" />
+    <h3 v-else-if="subscribed_course_events.length == 0 && !loading"
         class="d-flex justify-center subtitle">{{ $t('course_event.list.none_subscribed') }}</h3>
     <v-card v-else
             elevation="0"
@@ -42,6 +42,9 @@
     computed: {
       ...mapState('course_event', ['course_events']),
       ...mapState('layout', ['loading']),
+      subscribed_course_events() {
+        return this.course_events.filter(course_event => course_event.subscribed )
+      }
     },
     methods: {
      ...mapActions('course_event', ['fetchCourseEvents']),